contact routes are now protected

+++ controllers/contactController.js
@@ -5,7 +5,9 @@ const Contact = require("../models/contactModel");
 //@route GET /api/contacts
 //@access private
 const getContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({
+    user_id: req.user.id,
+  });
   res.status(200).json(contacts);
 });
 
@@ -24,6 +26,7 @@ const createContact = asyncHandler(async (req, res) => {
     name,
     email,
     phone,
+    user_id: req.user.id,
   });
 
   res.status(201).json(contact);
@@ -60,6 +63,11 @@ const updateContact = asyncHandler(async (req, res) => {
     throw new Error("Contact not found.");
   }
 
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("You are not authorized to update this contact.");
+  }
+
   const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
     req.body,
@@ -80,6 +88,11 @@ const deleteContact = asyncHandler(async (req, res) => {
   if (!contact) {
     res.status(404);
     throw new Error("Contact not found.");
+  }
+
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("You are not authorized to delete this contact.");
   }
 
   await Contact.deleteOne({ _id: req.params.id });
